Create 494. Target Sum
Solution represents the idea of subset sum problem. Array is divided into 2 parts: 1 contain all the +ve integers and other contains all the -ve integers, and we want in how many ways we can make a given target S.
Input: nums is [1, 1, 1, 1, 1], S is 3. 
Output: 5
Explanation: 

-1+1+1+1+1 = 3
+1-1+1+1+1 = 3
+1+1-1+1+1 = 3
+1+1+1-1+1 = 3
+1+1+1+1-1 = 3

There are 5 ways to assign symbols to make the sum of nums be target 3.